Initial commit
Created a lot of initial groundwork for the project. The actual library
made in zig is pretty bare bones, but the build script is fairly
complete and am happy with it.
The website has been setup with auth and a db connection but no actual
contrib implementations are done and there are missing sections. Not
too far from being done for the initial scope so just the usual web
shenanigans.